moved test to existing test_admin
